+++ EmotionCore/src/Primitives/Vector2.cs
@@ -156,7 +156,7 @@ namespace Emotion.Primitives
         /// <summary>
         /// Defines the size of the Vector2 struct in bytes.
         /// </summary>
-        public static readonly int SizeInBytes = Marshal.SizeOf(new Vector2());
+        public static readonly uint SizeInBytes = (uint) Marshal.SizeOf(new Vector2());
 
         /// <summary>
         /// Adds two vectors.
