+++ EmotionCore/src/Primitives/Vector3.cs
@@ -181,7 +181,7 @@ namespace Emotion.Primitives
         /// <summary>
         /// Defines the size of the Vector3 struct in bytes.
         /// </summary>
-        public static readonly int SizeInBytes = Marshal.SizeOf(new Vector3());
+        public static readonly uint SizeInBytes = (uint) Marshal.SizeOf(new Vector3());
 
         /// <summary>
         /// Adds two vectors.
